Switch from `globus task wait` to `globus task show`
We need to trap signals so that we can properly requeue ourselves.
The problem is, in order for BASH to properly trap a signal, it needs
to _not_ be having a command running.  In other words, if you are
executing a long-running command, you either need to have that long-
running command run in a background subshell, or you need to have the
command timeout occasionally.

`globus task wait` does have a --timeout option, but the exit code for
a timeout is 1, which is the same exit code for things like network
and server errors.  This makes it really hard to figure out why we
exited.

This switches us to use `globus task show`, with a JMESPath filter to
only output the status.  An exit code of 0 means we should check the
output for the job status, and then act accordingly.  A non-zero exit
code means a network or server timeout.  And since we're running this
in a loop with a BASH sleep, we can trap sigals with no problem (just
keep the `sleep` from being too long!).