Add the Flow Definition
This, combined with the Input Schema, is our Flow.  It has six major
steps:

1. **Compute Paths**: This is the `Compute_Paths` state.  This takes our
   inputs and computes all of the different paths.
2. **Make Directories**: This is `MakeDir_Compute`,
   `MakeDir_Compute_Input`, `MakeDir_Compute_Output`, and
   `MakeDir_Results`, to create directories for temporary storage and
   results storage.
3. **Copy Inputs**: This is `Copy_Inputs_To_Compute`, a simple Transfer.
4. **Compute**: This is `Do_Compute`, which calls Globus Compute to do
   the checksumming.
5. **Copy Results**: This is `Copy_Results_To_Dest`, another simple Transfer.
6. **Clean Up**: This is `Cleanup_Compute`, which is always run; and
   `Cleanup_Results`, which is only run if some part of the Flow fails.

Every state (except for the `End` states) catches exceptions,
transitioning to an approriate `Cleanup` state depending on the nature
of the exception.  The Flow ends in the `End` state, which goes to
either `EndFailure` or `EndSuccess`, depending on if there was a problem
with the flow.

WARNING: This Flow Definition is specific to the [Stanford SCG
Bioinformatics Cluster](https://login.scg.stanford.edu).  If you run
this elsewhere, you will need to change the Flow Definition.  In
particular, you will need to edit the `Compute_Paths` state: Change
`function_in_dir` and `function_out_dir` to the appropriate absolute
paths on your cluster (at least, in the part of the cluster where the
Compute Function will run).

For example, our Input Schema has the `cluster_temp` Collection pointing
to [SRCC SCG Lab
Storage](https://app.globus.org/file-manager/collections/3257fc54-9071-42fa-88ca-6097b2679b9a/overview).
The base path of that Collection is `/labs`, so the Flow Definition
prepends `/labs` when computing `function_in_dir` and
`function_out_dir`.