Add a script that runs & monitors the Flow
This (long) script provides a way to run and monitor the Flow without
needing to use the Globus web site.  It...

* Handles user login and authorization in a smart way: Ideally, you will
  only need to log in to Globus once, though you may need to log in
  twice.

* Pulls defaults from the Flow's Input Schema

* Presents a Click-based CLI for changing defaults, and capturing the
  Compute Endpoint & Function UUIDs.

* Validates Collections, Paths, the Compute Endpoint, and the Function.

* Starts the Flow

* Monitors the status of the Flow, displaying log entries similar to how
  the Globus web site displays log entries.

The script supports PEP 732 Inline Script metadata, so tools which
support that can run the script without the need to manually create a
virtual environment.

NOTE: The script cannot verify that the Function is compatible with the
Globus Compute Endpoint.  In general, they should be running the same
Python version.

WARNING: The script cannot verify that the Compute Collection & Path are
accessible by the Compute Endpoint & Function.  If you run this outside
of SCG, you need to change the Flow Definition and the Input Schema, and
possibly also this script!