Add the Function-registering code
This is the code which defines our Globus Compute Function, and which
also handles registering the Function with Globus Compute.

The script supports PEP 732 Inline Script metadata, so tools which
support that can run the script without the need to manually create a
virtual environment.

This uses the Globus Compute SDK: It saves us several steps that we
would need to do if we used the normal Globus SDK.  For example, we
don't need to create a Globus Auth Native Client.

The first thing the script does is define the Function.  We do this
before we import anything: When the function runs, it will not have
access to anything imported outside of the function.  Once the function
is defined, only then do we do global imports.

After the function is defined, the remaining code in the script handles
Function testing and registration.  The user is able to run the Function
locally, for testing & demonstration purposes.  The user is also able to
register the Function with Globus Compute (this part uses the Globus
Compute SDK).  The Globus Compute SDK handles all the work of user
authentication.

NOTE: In general, the Python version used to register the Function
should match the Python version used to run the Globus Compute Endpoint.