Start a README
The README includes an explanation of what this all is, and what it is
designed to accomplish.  Also included is a description of the
environment it is meant to work in, and the
copyright/license/contributions notice.

Still to come are instructions on how to set up and customize the
environment yourself.