README: Add info on the Flow Definition
This adds information on the Flow Definition.  It starts with a diagram of
the states and their transitions.

Next up comes a summary of the different types of states, and their
names.  An explanation is provided for how cleanup happens, and how
the Flow ends.

Finally, details are provided on the `Compute_Function_Arguments` state,
and how it will need to be customized when using the Flow on anything
other than SRCC SCG Lab Storage.