README: Add instructions on how to run the demo on SCG
The demo is easier to run on SCG, because the publicly-available Flow
Definition is configured to use the [SRCC SCG Lab
Storage](https://app.globus.org/file-manager/collections/3257fc54-9071-42fa-88ca-6097b2679b9a/overview)
Collection for cluster temp storage.

The instructions explain how to set up a Python virtual environment with
the correct packages, how to register the Compute Function, how to
configure and run the Compute Endpoint, and finally a demo of running
the Flow.

As part of writing this, I created asciicasts demonstrating each part of
the demo.  Those asciicasts are available on Asciinema, and are also
included in the repository.