README: Add a demo video
As the first thing in the README, include a link to a recording of the
demo that can be performed using this repository.

Since GitHub's display of READMEs does not really support dynamic
elements, I do not include a YouTube player.  Instead, I include a
thumbnail and link to the video.

Also, I include a link to the source file, which is on Google Drive.