run_and_monitor: Catch GARE errors
Many Globus operations can return GAREs: Globus Auth Requirements Errors.
These take the form of additional data within a GlobusAPIError, and are used to
indicate the need for additional authentication—above and beyond the normal
scope-based authentication—to meet some requirements of the resource server.

GAREs often appear when working with things that need more-frequent
authentication.  For example, if you run a demo with an HA Collection, it might
need re-authentication with MFA.

This commit adds code around the directory-stat and start-flow calls.  If a
GARE is returned, we run the user through an additional Globus Auth, then we
retry.